test(core): change fun arguments type

+++ src/test/kotlin/study/calculator/CalculatorTest.kt
@@ -9,22 +9,22 @@ class CalculatorTest {
 
     @Test
     fun `test addition`() {
-        assertThat(calculator.add(2, 3)).isEqualTo(5)
+        assertThat(calculator.add(2.toDouble(), 3.toDouble())).isEqualTo(5.toDouble())
     }
 
     @Test
     fun `test subtraction`() {
-        assertThat(calculator.subtract(2, 3)).isEqualTo(-1)
+        assertThat(calculator.subtract(2.toDouble(), 3.toDouble())).isEqualTo((-1).toDouble())
     }
 
     @Test
     fun `test multiplication`() {
-        assertThat(calculator.multiply(2, 3)).isEqualTo(6)
+        assertThat(calculator.multiply(2.toDouble(), 3.toDouble())).isEqualTo(6.toDouble())
     }
 
     @Test
     fun `test division`() {
-        assertThat(calculator.divide(4, 2)).isEqualTo(2)
+        assertThat(calculator.divide(4.toDouble(), 2.toDouble())).isEqualTo(2.toDouble())
     }
 
     @Test
@@ -47,6 +47,6 @@ class CalculatorTest {
 
     @Test
     fun `evaluates expression with all operators`() {
-        assertThat(calculator.evaluate("2 + 3 * 4 / 2")).isEqualTo(10)
+        assertThat(calculator.evaluate("2 + 3 * 4 / 2")).isEqualTo(10.toDouble())
     }
 }
